Update setup.py
Trying to work around SWIG didn't work. The _wrap.c file made with SWIG In OS X didn't work on a Linux install, so it seems that having SWIG installed is indeed necessary. For Linux the extra compile args needed to be commented out, and whether or not this breaks OS X is unclear.